pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dice-rl"
version = "0.1.0"
description = "Efficient RL Finetuning of Generative BC Policies"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "av==12.3.0",
    "einops==0.8.0",
    "gdown==5.2.0",
    "gym==0.22.0",
    "huggingface_hub[cli]==0.23.4",
    "hydra-core==1.3.2",
    "imageio==2.35.1",
    "matplotlib==3.7.5",
    "omegaconf==2.3.0",
    "pretty_errors==1.2.25",
    "torch==2.4.0",
    "tqdm==4.66.5",
    "wandb==0.17.7"
]

[project.optional-dependencies]
gym = [
    "cython<3",
    "d4rl",
    "patchelf",
]
robomimic = [
    "cython<3",
    "d4rl",
    "patchelf",
    "robomimic @ git+https://github.com/ARISE-Initiative/robomimic.git",
    "robosuite @ git+https://github.com/ARISE-Initiative/robosuite.git@v1.4.1",
]
exact = [   # exact likelihood
    "torchdiffeq",
]
all = [
    "cython<3",
    "d4rl",
    "gin-config",
    "ipdb==0.13.13",
    "patchelf",
    "pin",
    "robomimic @ git+https://github.com/ARISE-Initiative/robomimic.git",
    "robosuite @ git+https://github.com/ARISE-Initiative/robosuite.git@v1.4.1",
]

[project.urls]
Homepage = "https://diffusion-ppo.github.io/"

[tool.setuptools.packages.find]
exclude = []

[tool.wheel]
exclude = []
